Disable instant reset in ascent basement
Check with 'stage_flag' that the player is no longer in the basement. The main difference between basement and ascent basement is that the player must enter Depth2 before, this is done via 'visit_depths2'. Also, the game time must be checked to correctly set the flags back to nil (otherwise, you can press R twice to instant reset the game while in the ascent basement).
Proof of concept for main.lua code, typescript code was not touched.